+++ images/backend/src/index.js
@@ -2,6 +2,7 @@ import express from "express"
 import bodyParser from "body-parser"
 import cors from "cors"
 import { databaseService } from "./services/databaseService.js"
+import e from "express"
 
 const app = express()
 app.use(cors())
@@ -16,7 +17,7 @@ const port = 3000 || process.env.PORT
 app.listen(port, () => {
     console.log(`Example app listening at http://localhost:${port}`)
 });
- 
+
 /**
  * object template
  * 
@@ -35,6 +36,10 @@ app.listen(port, () => {
  *      @property {string} newEmail - new email of user
  *      @property {string} currentPassword - old password of user
  *      @property {string} newPassword - new password of user
+ * 
+ * @typedef {object} updatedTask
+ *      @property {string} currentTask - old task of user
+ *      @property {string} newTask - new task of user
  */
 
 /**GET endpoint, html page with all endpoints 
@@ -45,20 +50,24 @@ app.get('/', (request, response) => {
     response.status(200).redirect('./index.html')
 })
 
+
 /** POST endpoint, create a new user
  * 
  * @param {object} newUser - the new user to be created
  * @returns {string} - a message saying the user was created or an error message 
  */
 app.post('/createUser', async (request, response) => {
+    //check if all information is provided
     if (request.body.firstName && request.body.lastName && request.body.email && request.body.password) {
         try {
             const existingUserResult = await databaseService.getUser(request.body.email)
-            
+            //check if user already exists
             if (existingUserResult) {
+                //return error message if user already exists
                 response.status(400).send({
                     status: 400,
-                    message:'User already exists'})
+                    message: 'User already exists'
+                })
 
             } else {
                 const newUser = {
@@ -67,10 +76,12 @@ app.post('/createUser', async (request, response) => {
                     email: request.body.email,
                     password: request.body.password
                 }
+                //add user to database
                 await databaseService.addUser(newUser)
                 response.status(200).send({
                     status: 200,
-                    message:"User succesfully created"})
+                    message: "User succesfully created"
+                })
             }
         } catch (error) {
             //show error in console
@@ -83,10 +94,11 @@ app.post('/createUser', async (request, response) => {
         }
     } else {
         //return error message if missing information
-        response.status(400).send({
-            status: 400,
-            message:'Missing information'})
-        
+        response.status(401).send({
+            status: 401,
+            message: 'Missing information'
+        })
+
     }
 })
 
@@ -113,31 +125,90 @@ app.post('/login', async (request, response) => {
                 if (user.password === userCredentials.password) {
                     response.status(200).send({
                         status: 200,
-                        message:"User succesfully logged in"})
+                        message: "User succesfully logged in"
+                    })
                 } else {
-                    response.status(400).send({
-                        status: 400,
-                        message:"incorrect credentials"})
+                    response.status(401).send({
+                        status: 401,
+                        message: "incorrect credentials"
+                    })
                 }
             } else {
-                response.status(400).send({
-                    status: 400,
-                    message:"incorrect credentials"})
+                response.status(401).send({
+                    status: 401,
+                    message: "incorrect credentials"
+                })
             }
         } catch (error) {
             //show error in console
             console.log(error);
             //send error message in response
-            response.status(400).send({
-                status: 400,
-                message:error.message})
+            response.status(401).send({
+                status: 401,
+                message: error.message
+            })
         }
     } else {
-        response.status(400).send({
-            status: 400,
-            message:'incorrect credentials'})
+        response.status(401).send({
+            status: 401,
+            message: 'incorrect credentials'
+        })
     }
 })
+
+
+/** POST endpoint, get user data based on email
+ * 
+ * @param {object} userCredentials - user email and password
+ * @returns {object} User - user data
+ */
+app.post('/getUserData', async (request, response) => {
+    //check if email and password are provided
+    if (request.body.email && request.body.password) {
+        try {
+            const email = request.body.email;
+            // Assuming you have a function called getUserDataByEmail in your databaseService to fetch user data based on email
+            const userData = await databaseService.getUser(email);
+            //check if user exists and if password is correct
+            //if user does not exist, return error message
+            if (!userData) {
+                return response.status(404).json({
+                    status: 404,
+                    message: "User data not found"
+                });
+            }
+            //if password is incorrect, return error message
+            if (userData.password !== request.body.password) {
+                return response.status(401).json({
+                    status: 401,
+                    message: "Incorrect password"
+                });
+            } else {
+                //if user exists and password is correct, return user data
+                return response.status(200).json({
+                    status: 200,
+                    data: userData
+                });
+            }
+
+
+        } catch (error) {
+            // Handle the error here
+            console.error(error);
+            return response.status(500).json({
+                status: 500, error:
+                    "Internal server error"
+            });
+        }
+    } else {
+        //return error message if missing information
+        return response.status(400).json({
+            status: 400,
+            error: "Missing information"
+        });
+    }
+
+});
 
 /**PUT endpoint, update user credentials 
  * 
@@ -145,52 +216,58 @@ app.post('/login', async (request, response) => {
  * @returns {string} - a message saying the user was updated or an error message
 */
 app.put('/updateUser', async (request, response) => {
-   //get updated user credentials from request body
-   if (request.body.currentEmail && request.body.currentPassword) {
-    let updatedUser = {
-        currentEmail: request.body.currentEmail,
-        currentPassword: request.body.currentPassword,
-        newEmail: request.body.newEmail,
-        newPassword: request.body.newPassword
-    }
+    //get updated user credentials from request body
+    if (request.body.currentEmail && request.body.currentPassword) {
+        let updatedUser = {
+            currentEmail: request.body.currentEmail,
+            currentPassword: request.body.currentPassword,
+            newEmail: request.body.newEmail,
+            newPassword: request.body.newPassword
+        }
 
-    try {
-        let userCurrentCredentials = await databaseService.getUser(updatedUser.currentEmail)
-        //retrieve user from database
-        if (userCurrentCredentials) {
-            //check if current password is correct
-            if (userCurrentCredentials.password === updatedUser.currentPassword) {
-                //check if new email is not empty
-                databaseService.updateUser(updatedUser)
-                response.status(200).send({
-                    status: 400,
-                    message: "user succesfully updated"})
-            } else{
-                
+        try {
+            let userCurrentCredentials = await databaseService.getUser(updatedUser.currentEmail)
+            //retrieve user from database
+            if (userCurrentCredentials) {
+                //check if current password is correct
+                if (userCurrentCredentials.password === updatedUser.currentPassword) {
+                    //check if new email is not empty
+                    databaseService.updateUser(updatedUser)
+                    response.status(200).send({
+                        status: 200,
+                        message: "user succesfully updated"
+                    })
+                } else {
+
+                    response.status(401).send({
+                        status: 401,
+                        message: "incorrect password"
+                    })
+                }
+            } else {
                 response.status(401).send({
                     status: 401,
-                    message: "incorrect password"})
+                    message: "incorrect credentials"
+                })
             }
-        } else{
-            response.status(404).send({
-                status: 404,
-                message: "incorrect credentials"})
+        } catch (error) {
+            // Handle the error here
+            console.log(error);
+            //send error back
+            response.status(400).send({
+                status: 400,
+                message: error
+            })
         }
-    } catch (error) {
-        // Handle the error here
-        console.log(error);
-        //send error back
+    } else {
         response.status(400).send({
             status: 400,
-            message : error} )
+            message: "missing information"
+        })
     }
-   } else{
-         response.status(400).send({
-            status: 400,
-            message: "missing information"})
-   }
-    
 })
+
+
 
 /**DELETE endpoint, delete user from database
  * 
@@ -205,8 +282,10 @@ app.delete('/deleteUser', async (request, response) => {
                 email: request.body.email,
                 password: request.body.password
             }
+            console.log(userCredentials);
             //get user from database
             let user = await databaseService.getUser(userCredentials.email)
+            console.log(user);
             //check if user exists and if password is correct
             if (user) {
                 if (user.password === userCredentials.password) {
@@ -214,16 +293,21 @@ app.delete('/deleteUser', async (request, response) => {
                     await databaseService.deleteUser(userCredentials.email)
                     response.status(200).send({
                         status: 200,
-                        message:"User succesfully deleted"})
+                        message: "User succesfully deleted"
+                    })
                 } else {
-                    response.status(404).send({
-                        status: 404,
-                        message: "incorrect credentials"})
-                        }
+                    response.status(401).send({
+                        status: 401,
+                        message: "incorrect credentials"
+                    })
+                    console.log("log1");
+                }
             } else {
-                response.status(404).send({
-                        status: 404,
-                        message: "incorrect credentials"})
+                response.status(401).send({
+                    status: 401,
+                    message: "incorrect credentials"
+                })
+                console.log("log2");
             }
         } catch (error) {
             //show error in console
@@ -232,8 +316,221 @@ app.delete('/deleteUser', async (request, response) => {
             response.status(400).send(error.message)
         }
     } else {
-        response.status(404).send({
-                        status: 404,
-                        message: "incorrect credentials"})
-    }   
+        response.status(401).send({
+            status: 401,
+            message: "incorrect credentials"
+        })
+    }
+})
+
+/**POST endpoint, create new task 
+ * 
+ * @param {object} userCredentials, newTask - user credentials and new task
+*/
+app.post('/createTask', async (request, response) => {
+    //check if email and password are provided
+    if (request.body.email && request.body.password && request.body.newTask) {
+        try {
+            //get user from database
+            const result = await databaseService.getUser(request.body.email)
+            //check if user exists and if password is correct
+            if (result) {
+                //check if user exists and if password is correct
+                if (result.password === request.body.password) {
+                    //create new task
+                    await databaseService.createTask(result.id, request.body.newTask)
+                    response.status(200).send({
+                        status: 200,
+                        message: "task succesfully created"
+                    })
+                } else {
+                    //return error message if password is incorrect
+                    response.status(401).send({
+                        status: 401,
+                        message: "incorrect credentials"
+                    })
+                }
+            } else {
+                //return error message if user does not exist
+                response.status(401).send({
+                    status: 401,
+                    message: "missing details"
+                })
+            }
+        } catch (error) {
+            //show error in console
+            console.log(error);
+            //send error message in response
+            response.status(400).send(error.message)
+        }
+        
+            
+    } else {
+        response.status(401).send({
+            status: 401,
+            message: "incorrect credentials"
+        })
+    }
+})
+
+
+
+/**POST endpoint, get user tasks from database
+ * 
+ * @param {object} userCredentials - user credentials
+ * @returns {object} - user tasks
+*/
+app.post('/getTasks', async (request, response) => {
+    //check if email and password are provided
+    if (request.body.email && request.body.password) {
+        try {
+            //get user from database
+            const result = await databaseService.getUser(request.body.email)
+            //check if user exists and if password is correct
+            if (result) {
+                //check if user exists and if password is correct
+                if (result.password === request.body.password) {
+                    //get tasks from database
+                    const tasks = await databaseService.getTasks(result.id)
+                    response.status(200).send({
+                        status: 200,
+                        tasks: tasks
+                    })
+                } else {
+                    //return error message if password is incorrect
+                    response.status(401).send({
+                        status: 401,
+                        message: "incorrect credentials"
+                    })
+                }
+            } else {
+                //return error message if user does not exist
+                response.status(401).send({
+                    status: 401,
+                    message: "incorrect credentials"
+                })
+            }
+        } catch (error) {
+            //show error in console
+            console.log(error);
+            //send error message in response
+            response.status(400).send(error.message)
+        }
+        
+            
+    } else {
+        response.status(401).send({
+            status: 401,
+            message: "missing information"
+        })
+    }
+})
+
+
+/**PUT endpoint, update task
+ * 
+ * @param {object} userCredentials, updatedTask - user credentials and updated task
+ * @returns {string} - a message saying the task was updated or an error message
+*/
+app.put('/updateTask', async (request, response) => {
+    //check if email,password, currentTask and newTask are provided
+    if (request.body.email && request.body.password && request.body.currentTask && request.body.newTask) {
+        try {
+
+            //get user from database
+            const result = await databaseService.getUser(request.body.email)
+
+            //check if user exists and if password is correct
+            if (result) {
+                //check if user exists and if password is correct
+                if (result.password === request.body.password) {
+
+                    const updatedTask = {
+                        currentTask: request.body.currentTask,
+                        newTask: request.body.newTask
+                    }
+
+                    //update task in database
+                    await databaseService.updateTask(updatedTask)
+                    response.status(200).send({
+                        status: 200,
+                        message: "task succesfully updated"
+                    })
+                } else {
+                    //return error message if password is incorrect
+                    response.status(401).send({
+                        status: 401,
+                        message: "incorrect credentials"
+                    })
+                }
+
+            } else{
+                //return error message if user does not exist
+                response.status(401).send({
+                    status: 401,
+                    message: "incorrect credentials"
+                })
+            }
+        } catch (error) {
+            //show error in console
+            console.log(error);
+            //send error message in response
+            response.status(400).send(error.message)
+        }
+    } else {
+        response.status(400).send({
+            status: 400,
+            message: "missing information"
+        })
+    }
+})
+
+
+/**DELETE endpoint, delete task from database
+ * 
+ * @param {object} userCredentials, task - user credentials and task
+ * @returns {string} - a message saying the task was deleted or an error message
+*/
+app.delete('/deleteTask', async (request, response) => {
+    //check if email,password and task are provided
+    if (request.body.email && request.body.password && request.body.task) {
+        try {
+            //get user from database
+            const result = await databaseService.getUser(request.body.email)
+            //check if user exists and if password is correct
+            if (result) {
+                //check if user exists and if password is correct
+                if (result.password === request.body.password) {
+                    //delete task from database
+                    await databaseService.deleteTask(result.id,request.body.task)
+                    response.status(200).send({
+                        status: 200,
+                        message: "task succesfully deleted"
+                    })
+                } else {
+                    //return error message if password is incorrect
+                    response.status(401).send({
+                        status: 401,
+                        message: "incorrect credentials"
+                    })
+                }
+            } else {
+                //return error message if user does not exist
+                response.status(401).send({
+                    status: 401,
+                    message: "incorrect credentials"
+                })
+            }
+        } catch (error) {
+            //show error in console
+            console.log(error);
+            //send error message in response
+            response.status(400).send(error.message)
+        }
+    } else {
+        response.status(400).send({
+            status: 400,
+            message: "missing information"
+        })
+    }
 })